Create shared SecurityBenchmark creator function
