+++ Common/Brokerages/BitfinexBrokerageModel.cs
@@ -85,7 +85,7 @@ namespace QuantConnect.Brokerages
         public override IBenchmark GetBenchmark(SecurityManager securities)
         {
             var symbol = Symbol.Create("BTCUSD", SecurityType.Crypto, Market.Bitfinex);
-            return CreateSecurityBenchmark(securities, symbol);
+            return SecurityBenchmark.CreateInstance(securities, symbol);
         }
 
         /// <summary>
