+++ Common/Brokerages/GDAXBrokerageModel.cs
@@ -110,7 +110,7 @@ namespace QuantConnect.Brokerages
         public override IBenchmark GetBenchmark(SecurityManager securities)
         {
             var symbol = Symbol.Create("BTCUSD", SecurityType.Crypto, Market.GDAX);
-            return CreateSecurityBenchmark(securities, symbol);
+            return SecurityBenchmark.CreateInstance(securities, symbol);
         }
 
         /// <summary>
